+++ main.hpp
@@ -13,7 +13,7 @@
 struct index {
 	long long start = 0;
 	long long end = 0;
-	unsigned long long ix = 0;
+	__int128_t ix = 0;
 	short segment = 0;
 	unsigned short t : 2;
 	unsigned long long seq : (64 - 18);  // pack with segment and t to stay in 32 bytes
